refact(core): extract the schema sampler as an internal seam of OpenApiFixtureSource
OpenApiSchemaSampler owns schema-shaped knowledge — example-like
keywords, allOf/oneOf/anyOf composition, typed placeholders, and JSON
Pointer resolution — behind a one-method interface. It stays out of the
package barrel: the OpenAPI Source's interface is unchanged, but the
~150 pure sampling lines are now exercised with bare schema nodes
instead of a full spec literal per assertion.
