+++ packages/flutter_fixtures_core/lib/src/openapi_fixture_source.dart
@@ -4,6 +4,7 @@ import 'fixture_asset_loader.dart';
 import 'fixture_collection.dart';
 import 'fixture_document.dart';
 import 'http_fixture_source.dart';
+import 'openapi_schema_sampler.dart';
 
 /// Builds fixture collections from an OpenAPI 3.x JSON document.
 ///
@@ -25,6 +26,7 @@ class OpenApiFixtureSource implements HttpFixtureSource {
   final FixtureAssetLoader assetLoader;
 
   Map<String, dynamic>? _cachedSpec;
+  OpenApiSchemaSampler? _sampler;
 
   OpenApiFixtureSource({
     required this.specPath,
@@ -90,6 +92,9 @@ class OpenApiFixtureSource implements HttpFixtureSource {
     }
     return _cachedSpec = decoded.cast<String, dynamic>();
   }
+
+  OpenApiSchemaSampler _samplerFor(Map<String, dynamic> spec) =>
+      _sampler ??= OpenApiSchemaSampler(spec);
 
   /// The request paths to try: as given (query string dropped), plus with
   /// each `servers` base path stripped.
@@ -244,7 +249,7 @@ class OpenApiFixtureSource implements HttpFixtureSource {
       return [draft(fallbackIdentifier, mediaType['example'])];
     }
 
-    final sample = _sampleFromSchema(spec, mediaType['schema'], const {});
+    final sample = _samplerFor(spec).sample(mediaType['schema']);
     return [draft(fallbackIdentifier, sample)];
   }
 
@@ -273,156 +278,6 @@ class OpenApiFixtureSource implements HttpFixtureSource {
     return _deref(spec, value.cast<String, dynamic>());
   }
 
-  /// Generates sample data from a schema: explicit example-like values win,
-  /// then composition keywords, then a type-based placeholder. [seenRefs]
-  /// cuts `$ref` cycles (a cycle yields `null`).
-  Object? _sampleFromSchema(
-    Map<String, dynamic> spec,
-    Object? schemaNode,
-    Set<String> seenRefs,
-  ) {
-    if (schemaNode is! Map) {
-      return null;
-    }
-    var schema = schemaNode.cast<String, dynamic>();
-    final ref = schema[r'$ref'];
-    if (ref is String) {
-      if (seenRefs.contains(ref)) {
-        return null;
-      }
-      seenRefs = {...seenRefs, ref};
-      final resolved = _resolvePointer(spec, ref);
-      if (resolved is! Map) {
-        return null;
-      }
-      schema = resolved.cast<String, dynamic>();
-    }
-
-    if (schema.containsKey('example')) {
-      return schema['example'];
-    }
-    final examples = schema['examples'];
-    if (examples is List && examples.isNotEmpty) {
-      return examples.first;
-    }
-    if (schema.containsKey('const')) {
-      return schema['const'];
-    }
-    if (schema.containsKey('default')) {
-      return schema['default'];
-    }
-    final enumValues = schema['enum'];
-    if (enumValues is List && enumValues.isNotEmpty) {
-      return enumValues.first;
-    }
-
-    final allOf = schema['allOf'];
-    if (allOf is List && allOf.isNotEmpty) {
-      final merged = <String, dynamic>{};
-      Object? nonObject;
-      for (final sub in allOf) {
-        final sample = _sampleFromSchema(spec, sub, seenRefs);
-        if (sample is Map) {
-          merged.addAll(sample.cast<String, dynamic>());
-        } else {
-          nonObject ??= sample;
-        }
-      }
-      return merged.isNotEmpty ? merged : nonObject;
-    }
-    for (final keyword in const ['oneOf', 'anyOf']) {
-      final variants = schema[keyword];
-      if (variants is List) {
-        for (final sub in variants) {
-          final sample = _sampleFromSchema(spec, sub, seenRefs);
-          if (sample != null) {
-            return sample;
-          }
-        }
-      }
-    }
-
-    switch (_primaryType(schema)) {
-      case 'object':
-        return _objectSample(spec, schema, seenRefs);
-      case 'array':
-        final item = _sampleFromSchema(spec, schema['items'], seenRefs);
-        return item == null ? [] : [item];
-      case 'string':
-        return _stringSample(schema['format'] as String?);
-      case 'integer':
-        return 0;
-      case 'number':
-        return 0;
-      case 'boolean':
-        return true;
-      default:
-        return null;
-    }
-  }
-
-  Map<String, dynamic> _objectSample(
-    Map<String, dynamic> spec,
-    Map<String, dynamic> schema,
-    Set<String> seenRefs,
-  ) {
-    final result = <String, dynamic>{};
-    final properties = schema['properties'];
-    if (properties is Map) {
-      for (final entry in properties.entries) {
-        result[entry.key.toString()] =
-            _sampleFromSchema(spec, entry.value, seenRefs);
-      }
-    }
-    return result;
-  }
-
-  String _stringSample(String? format) {
-    switch (format) {
-      case 'date-time':
-        return '2024-01-01T12:00:00Z';
-      case 'date':
-        return '2024-01-01';
-      case 'time':
-        return '12:00:00';
-      case 'email':
-        return 'user@example.com';
-      case 'uuid':
-        return '00000000-0000-4000-8000-000000000000';
-      case 'uri':
-      case 'url':
-        return 'https://example.com';
-      case 'hostname':
-        return 'example.com';
-      case 'ipv4':
-        return '192.168.0.1';
-      default:
-        return 'string';
-    }
-  }
-
-  String? _primaryType(Map<String, dynamic> schema) {
-    final type = schema['type'];
-    if (type is String) {
-      return type;
-    }
-    // OpenAPI 3.1 allows a type array, e.g. ["string", "null"].
-    if (type is List) {
-      for (final t in type) {
-        if (t is String && t != 'null') {
-          return t;
-        }
-      }
-    }
-    if (schema['properties'] is Map) {
-      return 'object';
-    }
-    if (schema['items'] is Map) {
-      return 'array';
-    }
-    return null;
-  }
-
   /// Follows document-local `$ref`s on [node] until a concrete object is
   /// reached.
   Map<String, dynamic> _deref(
@@ -436,33 +291,12 @@ class OpenApiFixtureSource implements HttpFixtureSource {
       if (ref is! String || !seen.add(ref)) {
         return current;
       }
-      final resolved = _resolvePointer(spec, ref);
+      final resolved = _samplerFor(spec).resolvePointer(ref);
       if (resolved is! Map) {
         return current;
       }
       current = resolved.cast<String, dynamic>();
     }
-  }
-
-  Object? _resolvePointer(Map<String, dynamic> spec, String ref) {
-    if (!ref.startsWith('#/')) {
-      return null;
-    }
-    Object? node = spec;
-    for (final rawKey in ref.substring(2).split('/')) {
-      final key = rawKey.replaceAll('~1', '/').replaceAll('~0', '~');
-      if (node is Map) {
-        node = node[key];
-      } else if (node is List) {
-        final index = int.tryParse(key);
-        node = (index != null && index >= 0 && index < node.length)
-            ? node[index]
-            : null;
-      } else {
-        return null;
-      }
-    }
-    return node;
   }
 
   /// Suffixes repeated identifiers (`"Error (2)"`) so every document stays
